# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for SMM-auto application",
        "title": "SMM-Auto API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/bot/sendAudio": {
            "post": {
                "description": "Send text message to some channel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send audio",
                "operationId": "SendAudio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text of post",
                        "name": "caption",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token of bot that manages given channels",
                        "name": "botToken",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "channelName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Audio message file",
                        "name": "audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/bot/sendLocation": {
            "post": {
                "description": "Send location message to some channel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send location",
                "operationId": "SendLocation",
                "parameters": [
                    {
                        "description": "Location body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SendLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/bot/sendMessage": {
            "post": {
                "description": "Send text message to some channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send message",
                "operationId": "SendMessage",
                "parameters": [
                    {
                        "description": "Message body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "404": {
                        "description": "Channel with given ID does not exist",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "417": {
                        "description": "Error identifying user",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/bot/sendPhoto": {
            "post": {
                "description": "Send mediagroup message to some channel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send media",
                "operationId": "SendMediaGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text of post",
                        "name": "caption",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Media message file (max of 10 files are allowed)",
                        "name": "media",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of channel",
                        "name": "chat",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stringified map of each file type (Example: \\",
                        "name": "fileTypes",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/bot/sendVenue": {
            "post": {
                "description": "Send venue",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send venue",
                "operationId": "SendVenue",
                "parameters": [
                    {
                        "description": "Venue request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.SendVenueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/bot/sendVideo": {
            "post": {
                "description": "Send video message to some channel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send video",
                "operationId": "SendVideo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text of post",
                        "name": "caption",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Voice message file",
                        "name": "video",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "channelName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token of bot that manages given channels",
                        "name": "botToken",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
        "/bot/sendVideoNote": {
            "post": {
                "description": "Send video message to some channel",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bot"
                ],
                "summary": "Send video",
                "operationId": "SendVideoNote",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Voice message file",
                        "name": "videoNote",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "channelName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/jsonHelper.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.SendLocationRequest": {
            "type": "object",
            "properties": {
                "botToken": {
                    "type": "string"
                },
                "channelName": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                }
            }
        },
        "controllers.SendMessageRequest": {
            "type": "object",
            "properties": {
                "botToken": {
                    "type": "string"
                },
                "channelName": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "controllers.SendVenueRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "botToken": {
                    "type": "string"
                },
                "channelName": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "jsonHelper.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}